Clean up old buttons
