+++ packages/libs/react-ui/src/components/Button/Button.stories.tsx
@@ -1,19 +1,25 @@
-import { SystemIcon } from '@components/Icon';
-import { withCenteredStory } from '@storyDecorators';
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryFn, StoryObj } from '@storybook/react';
 import React from 'react';
+import { LeadingIcon, Plus, TrailingIcon } from '../Icon/System/SystemIcon';
+import { Box } from '../Layout/Box/Box';
+import { Heading } from '../Typography/Heading/Heading';
 import type { IButtonProps } from './Button';
 import { Button } from './Button';
-import { colorVariants, typeVariants } from './Button.css';
+import { button } from './SharedButton.css';
 
-const meta: Meta<
-  {
-    text: string;
-  } & IButtonProps
-> = {
+// eslint-disable-next-line @kadena-dev/typedef-var
+const buttonVariants = Object.keys(
+  (button as any).classNames?.variants?.variant,
+) as IButtonProps['variant'][];
+
+// eslint-disable-next-line @kadena-dev/typedef-var
+const buttonColors = Object.keys(
+  (button as any).classNames?.variants?.color,
+) as IButtonProps['color'][];
+
+const meta: Meta<IButtonProps> = {
   title: 'Components/Button',
   component: Button,
-  decorators: [withCenteredStory],
   parameters: {
     status: { type: 'inDevelopment' },
     controls: {
@@ -23,207 +29,217 @@ const meta: Meta<
     docs: {
       description: {
         component:
-          'The Button component renders a clickable element that can either be a button or anchor which will be styled according to the variant prop (`primary` being the default).<br /><br />The Button component can include an icon<sup>*</sup> which can be aligned either left or right (default: `right`).<br /><br /><em><sup>*</sup> Please use IconButton when you require a button with only an icon.</em>',
+          'The Button component renders a button which will be styled according to the variant/color prop (`contained/primary` being the default)',
       },
     },
   },
   argTypes: {
     onClick: {
       action: 'clicked',
-      if: { arg: 'as', eq: 'button' },
+      description: '(deprecated) callback when button is clicked',
       table: {
         disable: true,
       },
     },
-    icon: {
-      options: [
-        ...['-'],
-        ...Object.keys(SystemIcon),
-      ] as (keyof typeof SystemIcon)[],
-      control: {
-        type: 'select',
-      },
-      table: {
-        type: { summary: Object.keys(SystemIcon).join(' | ') },
-      },
-      if: { arg: 'loading', eq: false },
-    },
-    iconAlign: {
-      description: 'align icon to left or right',
-      options: ['left', 'right'] as IButtonProps['iconAlign'][],
-      control: {
-        type: 'radio',
-      },
-      table: {
-        type: { summary: 'left | right' },
-        defaultValue: { summary: 'right' },
-      },
-      if: { arg: 'icon', neq: '-' },
-    },
-    color: {
-      options: Object.keys(colorVariants) as (keyof typeof colorVariants)[],
-      control: {
-        type: 'select',
-      },
-      description: 'color variant',
-      table: {
-        type: { summary: Object.keys(colorVariants).join(' | ') },
-        defaultValue: { summary: 'primary' },
-      },
-    },
     variant: {
-      options: Object.keys(typeVariants) as (keyof typeof typeVariants)[],
+      options: buttonVariants,
       control: {
         type: 'select',
       },
       description: 'button style variant',
       table: {
-        type: { summary: Object.keys(typeVariants).join(' | ') },
+        type: { summary: buttonVariants.join(' | ') },
         defaultValue: { summary: 'default' },
       },
     },
-    title: {
-      control: {
-        type: 'text',
-      },
-      description: 'aria label',
-      table: {
-        type: { summary: 'string' },
-        defaultValue: { summary: '' },
-      },
-    },
-    text: {
-      description: 'label text',
-      control: {
-        type: 'text',
-      },
-    },
-    href: {
-      description: 'href is required when rendered as anchor',
-      control: {
-        type: 'text',
-      },
-      if: { arg: 'as', eq: 'a' },
-    },
-    target: {
-      description: 'only used when rendered as anchor',
-      options: ['_blank', '_self'] as IButtonProps['target'][],
-      control: {
-        type: 'radio',
-      },
-      if: { arg: 'as', eq: 'a' },
-    },
-    type: {
-      description: 'type of button',
-      options: ['button', 'submit', 'reset'] as IButtonProps['type'][],
+    color: {
+      options: buttonColors,
       control: {
         type: 'select',
       },
+      description: 'button color variant',
       table: {
-        type: { summary: 'button | submit | reset' },
-        defaultValue: { summary: 'button' },
-      },
-      if: { arg: 'as', eq: 'button' },
-    },
-    as: {
-      description: 'render as button or anchor',
-      options: ['button', 'a'] as IButtonProps['as'][],
-      control: {
-        type: 'radio',
-      },
-      table: {
-        type: { summary: 'button | a' },
-        defaultValue: { summary: 'button' },
+        type: { summary: buttonColors.join(' | ') },
+        defaultValue: { summary: 'default' },
       },
     },
-    disabled: {
+    isDisabled: {
       description: 'only used when rendered as button',
       control: {
         type: 'boolean',
       },
-      if: { arg: 'as', eq: 'button' },
     },
-    loading: {
+    isLoading: {
       description: 'loading state',
       control: {
         type: 'boolean',
       },
-      if: { arg: 'as', eq: 'button' },
     },
-    active: {
-      description: 'set to apply active visual state',
+    isCompact: {
+      description: 'compact button style',
       control: {
         type: 'boolean',
-        defaultValue: { summary: false },
       },
-    },
-    asChild: {
-      description:
-        'Allow users to pass on styles, icons, and additional props to the child component. For example when using next/link in Next.js.',
     },
   },
 };
 
-type Story = StoryObj<
+type ButtonStory = StoryObj<
   {
     text: string;
   } & IButtonProps
 >;
 
-export const Dynamic: Story = {
+export const _Button: ButtonStory = {
   name: 'Button',
   args: {
-    active: false,
-    as: 'button',
-    color: 'primary',
-    disabled: false,
-    href: 'https://kadena.io',
-    icon: undefined,
-    iconAlign: 'right',
-    loading: false,
-    target: '_self',
     text: 'Click me',
-    title: 'test title',
-    variant: 'default',
+    variant: 'contained',
+    color: 'primary',
+    isDisabled: false,
+    isCompact: false,
+    isLoading: false,
+    icon: undefined,
+    startIcon: undefined,
+    endIcon: undefined,
   },
-  render: ({
-    active,
-    as,
-    color = 'primary',
-    disabled,
-    href,
-    icon,
-    iconAlign = 'right',
-    loading,
-    onClick,
-    target,
-    text,
-    title,
-    variant = 'default',
-  }) => {
-    if (loading) {
-      icon = 'Loading';
-    }
-
-    return (
-      <Button
-        active={active}
-        as={as}
-        color={color}
-        disabled={disabled}
-        href={href}
-        icon={icon}
-        iconAlign={iconAlign}
-        loading={loading}
-        onClick={onClick}
-        target={target}
-        title={title}
-        variant={variant}
-      >
-        {text}
-      </Button>
-    );
+  render: ({ text, ...props }) => {
+    return <Button {...props}>{text}</Button>;
   },
 };
+
+export const AllVariants: StoryFn<IButtonProps> = ({
+  isCompact,
+  isDisabled,
+  isLoading,
+}) => (
+  <Box gap="xs" display="flex">
+    <Box gap="xs" display="flex" flexDirection="column" alignItems="flex-start">
+      <Heading variant="h6">Default</Heading>
+      {buttonColors.map((color) => (
+        <Button
+          key={color}
+          color={color}
+          isCompact={isCompact}
+          isDisabled={isDisabled}
+          isLoading={isLoading}
+          startIcon={<LeadingIcon />}
+          endIcon={<TrailingIcon />}
+        >
+          {color}
+        </Button>
+      ))}
+    </Box>
+
+    <Box gap="xs" display="flex" flexDirection="column" alignItems="flex-start">
+      <Heading variant="h6">Alternative</Heading>
+      {buttonColors.map((color) => (
+        <Button
+          key={color}
+          color={color}
+          variant="alternative"
+          isCompact={isCompact}
+          isDisabled={isDisabled}
+          isLoading={isLoading}
+          startIcon={<LeadingIcon />}
+          endIcon={<TrailingIcon />}
+        >
+          {color}
+        </Button>
+      ))}
+    </Box>
+
+    <Box gap="xs" display="flex" flexDirection="column" alignItems="flex-start">
+      <Heading variant="h6">Outlined</Heading>
+      {buttonColors.map((color) => (
+        <Button
+          key={color}
+          color={color}
+          variant="outlined"
+          isCompact={isCompact}
+          isDisabled={isDisabled}
+          isLoading={isLoading}
+          startIcon={<LeadingIcon />}
+          endIcon={<TrailingIcon />}
+        >
+          {color}
+        </Button>
+      ))}
+    </Box>
+
+    <Box gap="xs" display="flex" flexDirection="column" alignItems="flex-start">
+      <Heading variant="h6">Text</Heading>
+      {buttonColors.map((color) => (
+        <Button
+          key={color}
+          color={color}
+          variant="text"
+          isCompact={isCompact}
+          isDisabled={isDisabled}
+          isLoading={isLoading}
+          startIcon={<LeadingIcon />}
+          endIcon={<TrailingIcon />}
+        >
+          {color}
+        </Button>
+      ))}
+    </Box>
+  </Box>
+);
+
+export const StartIcon: StoryFn<IButtonProps> = ({
+  isCompact,
+  isDisabled,
+  isLoading,
+  color,
+  variant,
+}) => (
+  <Button
+    startIcon={<Plus />}
+    isCompact={isCompact}
+    isDisabled={isDisabled}
+    isLoading={isLoading}
+    color={color}
+    variant={variant}
+  >
+    Click me
+  </Button>
+);
+
+export const EndIcon: StoryFn<IButtonProps> = ({
+  isCompact,
+  isDisabled,
+  isLoading,
+  color,
+  variant,
+}) => (
+  <Button
+    endIcon={<Plus />}
+    isCompact={isCompact}
+    isDisabled={isDisabled}
+    isLoading={isLoading}
+    color={color}
+    variant={variant}
+  >
+    Click me
+  </Button>
+);
+
+export const OnlyIcon: StoryFn<IButtonProps> = ({
+  isCompact,
+  isDisabled,
+  isLoading,
+  color,
+  variant,
+}) => (
+  <Button
+    icon={<Plus />}
+    isCompact={isCompact}
+    isDisabled={isDisabled}
+    isLoading={isLoading}
+    color={color}
+    variant={variant}
+  />
+);
 
 export default meta;
