+++ packages/libs/react-ui/src/components/Button/index.ts
@@ -1,0 +1,1 @@
+export { Button, type IButtonProps } from './Button';
