--- packages/libs/react-ui/src/components/Button/index.tsx
@@ -1,2 +1,0 @@
-export { Button, type IButtonProps } from './Button';
-export { LinkButton, type ILinkButtonProps } from './LinkButton';
