--- packages/libs/react-ui/src/components/Link/Link.css.ts
@@ -1,28 +1,0 @@
-import { sprinkles } from '@theme/sprinkles.css';
-import { vars } from '@theme/vars.css';
-import { style } from '@vanilla-extract/css';
-
-export const linkContainerClass = style([
-  sprinkles({
-    display: 'inline-flex',
-    gap: '$2',
-    color: '$primaryContrastInverted',
-  }),
-  {
-    selectors: {
-      '&:hover': {
-        textDecoration: 'none',
-      },
-      '&:active': {
-        color: vars.colors.$negativeContrastInverted,
-      },
-      '&:visited': {
-        color: vars.colors.$tertiaryContrastInverted,
-      },
-    },
-  },
-]);
-
-export const blockLinkClass = style({
-  display: 'flex',
-});
